+++ lampray-web/src/main/java/tech/lamprism/lampray/system/database/CertificateValue.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package tech.lamprism.lampray.web.configuration.database
+package tech.lamprism.lampray.system.database
 
 import java.io.File
