+++ lampray-web/src/main/java/tech/lamprism/lampray/system/database/builders/DatabaseUrlBuilders.kt
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package tech.lamprism.lampray.web.configuration.database.builders
+package tech.lamprism.lampray.system.database.builders
 
-import tech.lamprism.lampray.web.configuration.database.DatabaseUrlBuilder
+import tech.lamprism.lampray.system.database.DatabaseUrlBuilder
 
 /**
  * Registry for all database URL builders.
