+++ lampray-web/src/main/java/tech/lamprism/lampray/system/database/builders/SQLiteUrlBuilder.kt
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package tech.lamprism.lampray.web.configuration.database.builders
+package tech.lamprism.lampray.system.database.builders
 
-import tech.lamprism.lampray.web.configuration.database.DatabaseConfig
-import tech.lamprism.lampray.web.configuration.database.DatabaseType
 import java.io.File
+import tech.lamprism.lampray.system.database.DatabaseConfig
+import tech.lamprism.lampray.system.database.DatabaseType
 
 /**
  * URL builder for SQLite databases.
